+++ src/game/game_fn_80031948.c
@@ -26,7 +26,7 @@ typedef struct EffectDescriptor {
 
 extern void fn_80181F5C(EffectDescriptor*);
 extern int fn_800FBFB0(void);
-extern u16 fn_801D3A24(s32, s32);
+extern s32 fn_801D3A24(s32, s32);
 extern void fn_80182014(void);
 extern void fn_80147EC4(EffectDescriptor*);
